+++ jaorm-core/src/main/java/io/github/ulisse1996/jaorm/specialization/SingleKeyDao.java
@@ -1,0 +1,12 @@
+package io.github.ulisse1996.jaorm.specialization;
+
+import io.github.ulisse1996.jaorm.BaseDao;
+
+import java.util.Optional;
+
+public interface SingleKeyDao<T, R> extends BaseDao<T> {
+
+    T readByKey(R key);
+    Optional<T> readOptByKey(R key);
+    int deleteByKey(R key);
+}