+++ jaorm-processor/src/main/java/io/github/ulisse1996/jaorm/processor/validation/impl/QueryValidator.java
@@ -1,19 +1,33 @@
 package io.github.ulisse1996.jaorm.processor.validation.impl;
 
+import io.github.ulisse1996.jaorm.annotation.Dao;
+import io.github.ulisse1996.jaorm.annotation.Id;
 import io.github.ulisse1996.jaorm.annotation.Query;
 import io.github.ulisse1996.jaorm.processor.exception.ProcessorException;
 import io.github.ulisse1996.jaorm.processor.strategy.QueryStrategy;
 import io.github.ulisse1996.jaorm.processor.util.ProcessorUtils;
 import io.github.ulisse1996.jaorm.processor.validation.Validator;
+import io.github.ulisse1996.jaorm.specialization.DoubleKeyDao;
+import io.github.ulisse1996.jaorm.specialization.SingleKeyDao;
+import io.github.ulisse1996.jaorm.specialization.TripleKeyDao;
 
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ExecutableElement;
+import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.TypeKind;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 public class QueryValidator extends Validator {
+
+    private static final List<SpecializationSpecific> SPECIALIZED_DAOS = Arrays.asList(
+            new SpecializationSpecific(SingleKeyDao.class, 1),
+            new SpecializationSpecific(DoubleKeyDao.class, 2),
+            new SpecializationSpecific(TripleKeyDao.class, 3)
+    );
 
     public QueryValidator(ProcessingEnvironment processingEnvironment) {
         super(processingEnvironment);
@@ -26,6 +40,40 @@ public class QueryValidator extends Validator {
                 .map(ExecutableElement.class::cast)
                 .collect(Collectors.toList());
         queries.forEach(this::checkQuery);
+        annotated.stream()
+                .filter(ele -> ele.getAnnotation(Dao.class) != null)
+                .map(TypeElement.class::cast)
+                .forEach(this::checkKeys);
+    }
+
+    private void checkKeys(TypeElement type) {
+        debugMessage("Check number of keys for Query " + type.getSimpleName());
+        for (SpecializationSpecific specific : SPECIALIZED_DAOS) {
+            Optional<TypeElement> entity = findOptSpecific(type, specific);
+            if (entity.isPresent()) {
+                long keys = findKeysNumber(entity.get());
+                if (keys != specific.parameters) {
+                    throw new ProcessorException(String.format("Error on %s ! Required %d @Id but found %d in Entity",
+                            type, specific.parameters, keys));
+                }
+            }
+        }
+    }
+
+    private long findKeysNumber(TypeElement typeElement) {
+        return ProcessorUtils.getAllValidElements(processingEnvironment, typeElement)
+                .stream()
+                .filter(el -> el.getAnnotation(Id.class) != null)
+                .count();
+    }
+
+    private Optional<TypeElement> findOptSpecific(TypeElement type, SpecializationSpecific specific) {
+        if (ProcessorUtils.isSubType(this.processingEnvironment, type, specific.klass)) {
+            String t = ProcessorUtils.getBaseDaoGeneric(processingEnvironment, type);
+            return Optional.of(this.processingEnvironment.getElementUtils().getTypeElement(t));
+        }
+
+        return Optional.empty();
     }
 
     private void checkQuery(ExecutableElement executableElement) {
@@ -72,6 +120,17 @@ public class QueryValidator extends Validator {
     private void assertVoid(ExecutableElement method) {
         if (!method.getReturnType().getKind().equals(TypeKind.VOID)) {
             throw new ProcessorException("Can't use Delete or Update statement with a non-void method");
+        }
+    }
+
+    private static class SpecializationSpecific {
+
+        private final int parameters;
+        private final Class<?> klass;
+
+        private SpecializationSpecific(Class<?> klass, int parameters) {
+            this.klass = klass;
+            this.parameters = parameters;
         }
     }
 }