+++ jaorm-test/src/test/java/io/github/ulisse1996/jaorm/integration/test/entity/DoubleIdEntity.java
@@ -1,0 +1,33 @@
+package io.github.ulisse1996.jaorm.integration.test.entity;
+
+import io.github.ulisse1996.jaorm.annotation.Column;
+import io.github.ulisse1996.jaorm.annotation.Id;
+import io.github.ulisse1996.jaorm.annotation.Table;
+
+@Table(name = "DOUBLE_ID")
+public class DoubleIdEntity {
+
+    @Id
+    @Column(name = "FIRST_ID")
+    private long firstId;
+
+    @Id
+    @Column(name = "SECOND_ID")
+    private long secondId;
+
+    public long getFirstId() {
+        return firstId;
+    }
+
+    public void setFirstId(long firstId) {
+        this.firstId = firstId;
+    }
+
+    public long getSecondId() {
+        return secondId;
+    }
+
+    public void setSecondId(long secondId) {
+        this.secondId = secondId;
+    }
+}