+++ jaorm-test/src/test/java/io/github/ulisse1996/jaorm/integration/test/query/CustomDoubleKeyDao.java
@@ -1,0 +1,10 @@
+package io.github.ulisse1996.jaorm.integration.test.query;
+
+import io.github.ulisse1996.jaorm.annotation.Dao;
+import io.github.ulisse1996.jaorm.integration.test.entity.DoubleIdEntity;
+import io.github.ulisse1996.jaorm.specialization.DoubleKeyDao;
+
+@Dao
+public interface CustomDoubleKeyDao extends DoubleKeyDao<DoubleIdEntity, Long, Long> {
+}
+