+++ jaorm-test/src/test/java/io/github/ulisse1996/jaorm/integration/test/query/SimpleUserDao.java
@@ -1,0 +1,9 @@
+package io.github.ulisse1996.jaorm.integration.test.query;
+
+import io.github.ulisse1996.jaorm.annotation.Dao;
+import io.github.ulisse1996.jaorm.integration.test.entity.User;
+import io.github.ulisse1996.jaorm.specialization.SingleKeyDao;
+
+@Dao
+public interface SimpleUserDao extends SingleKeyDao<User, Integer> {
+}